my understanding of CRPS
In the EXX terms, X and X' should two IID copies from posterior predictions, not two predictions conditional on the SAME theta. (Think about paired t test )

I add this random shuffle to decouple x1 and x2.  In essence, EXX is a double expectations, and requires PSIS of the products.